fix not finding any drives

+++ Azure-ARM/VM-Extension-Scripts/Format-AllDrives.ps1
@@ -1,8 +1,13 @@
 $i = 0;
-(get-disk).where({$_.OperationalStatus -eq "Offline"}) |% {
+(get-disk).where({$_.Number -notin "0","1"}) |% {
     $d = $_;
+    if($d.PartitionStyle -ne "RAW"){
+        continue;
+    }
     Initialize-Disk -Number $($d.Number) -PartitionStyle GPT;
     $dp = New-Partition -DiskNumber $($d.Number) -UseMaximumSize -AssignDriveLetter;
     Format-Volume -DriveLetter $($dp.DriveLetter) -NewFileSystemLabel "Service-Fabric-$i" -FileSystem NTFS;
     $i++
 }
+
+exit 0;
